+++ src/main/java/insane96mcp/survivalreimagined/module/mining/solarium/item/SolariumArmorItem.java
@@ -29,4 +29,8 @@ public class SolariumArmorItem extends ArmorItem {
 		super.inventoryTick(pStack, pLevel, pEntity, pSlotId, pIsSelected);
 		Solarium.healGear(pStack, pEntity, pLevel);
 	}
+
+	public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+		return !ItemStack.isSameItem(oldStack, newStack);
+	}
 }