+++ src/main/java/insane96mcp/survivalreimagined/module/mining/solarium/item/SolariumSwordItem.java
@@ -1,11 +1,7 @@
 package insane96mcp.survivalreimagined.module.mining.solarium.item;
 
-import com.google.common.collect.Multimap;
 import insane96mcp.survivalreimagined.module.mining.solarium.Solarium;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.ai.attributes.Attribute;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SwordItem;
 import net.minecraft.world.level.Level;
@@ -21,8 +17,7 @@ public class SolariumSwordItem extends SwordItem {
 		Solarium.healGear(pStack, pEntity, pLevel);
 	}
 
-	@Override
-	public Multimap<Attribute, AttributeModifier> getAttributeModifiers(EquipmentSlot slot, ItemStack stack) {
-		return super.getAttributeModifiers(slot, stack);
+	public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+		return !ItemStack.isSameItem(oldStack, newStack);
 	}
 }